Fix tab selector in Hockey Statsheet
The jquery selector used to set the active tab on page load was breaking due to the presence of un-stringified hash (#) character.  This was causing the subsequent failure of the team-specific player selectors when entering goals and penalties, essentially causing players from both team to be visible at all times.  If a user were to enter a goal or penalty and select players by keying the player number the selection may result with the player from the wrong team, which is how this issue was discovered and reported