Account Confirmation period of 2 days #104
with this change we reconfigure the devise user management system to allow new user accounts to be used immediately.  An email confirmation message is still sent to the user and the account must be confirmed, however the user will be able to use the system unconfirmed for a period of 2 days.  In the future we may need to add special confirmation checks so that senstive features cannot be accessed within the confirmation grace period.