verify_admin - determine whether user is logged in
Problem: the verify_admin filter assumed the user was not an admin because he/she was not logged in and thus when one was not an admin the response was to redirect the user to the signin page.  If the user was logged in then when landing on the signin page he/she would be redirected back to the prior page... which would then verify_admin and redirect back to the signin page... repeat, repeat, repeat... problem!

Solution: verify_admin now raises a 401-not authorized error page for logged in users who attempt to access a page that they do not have permission to view

fixes #116