Remove hard-coded ical timezone adjustment
sportified originally persisted event starts_on date/time values into the UTC timezone in order to avoid complications of time zones... yet when event information was passed to systems that were savvy to timezones things broke down.  For example, when passing a list of events/games to a mobile device's calendar the mobile device would compensate for time zone and our events would be off.  In order to make this work for mobile devices and calendar software we had hard-coded an adjustment to make the event appear at the proper time... however recently we have implemented correct timezone settings into the application and our prior correction (err... hack) not makes the correctly-time-zoned events appear 7 hours off.  With this commit we remove the previous hard-coded correction.

also - have a funny feeling that this hack exists somewhere else and will need to be removed from there as well